Moves airtable files from app/common to app/common/airtable
